Unlink entities during PreUnload
This fixes a bug where reloading the plugin would destroy the entity components asynchronously during unload, after the newly loaded instance of the plugin had decided not to create new entity components due to the presence of the existing ones. Practically, this bug was causing the debug info to not show, and commands to not recognize the entities.